edusweep: Name quarantine files using GUID
This change attempts to prevent file naming conflicts by using the GUID
for the name of the file itself in quarantine, as well as the JSON descriptor.

Backwards compatibility is maintained by using an on-demand property:
OriginalName to show the original name of the file in the quarantine
manager list.